restructure directory, add jsdoc, refractor code
