+++ README.md
@@ -73,8 +73,11 @@ chmod a+x cron.sh
 ## Contributing
 Contributions are welcome, please look in the issues tab to contribute.
 *   VSCode is recommended as the editor for this project.
-*   Install eslint plugin for vscode.
-*   Use this [video][1] to setup eslint-prettier for vscode.
+*   Install eslint plugin for vscode. Use this [video][1] to setup eslint-prettier for vscode.
+*   Please follow [JSDoc][3] for your submitted code. Use [DocumentThis][2] plugin      if required.
 *   Please read CONTRIBUTING.md guide to know more.
+*   Write tests for complicated components and modules.
 
 [1]: https://www.youtube.com/watch?v=YIvjKId9m2c
+[2]: https://marketplace.visualstudio.com/items?itemName=joelday.docthis
+[3]: http://usejsdoc.org/
