+++ lib/cli/questions_cli.js
@@ -2,7 +2,7 @@
 const chalk = require("chalk");
 const path = require("path");
 
-const PATH_TO_DOTENV = path.resolve(__dirname, ".env");
+const PATH_TO_DOTENV = path.resolve(__dirname, "../../.env");
 
 require("dotenv").config({
   path: PATH_TO_DOTENV
