+++ lib/email_notification.js
