+++ lib/resolve_raw_data.js
@@ -1,11 +1,18 @@
 /* eslint-disable camelcase */
+
 const TO_REMOVE = "View Friendship";
 const TO_MATCH = ['class="_tzn"><a', '_ohe"><a'];
 
+/**
+ * Retrieves firstNames and fullNames from the raw_names array 
+ *
+ * @param {*} raw_names - array of names of birthday profiles in raw format
+ * @returns an object that contains an array of firstNames and fullNames
+ */
 function resolveNames(raw_names) {
   const fullNames = raw_names.split("\n\n");
   /* for (let i = fullNames.length; i--; ) */
-  for (let i = fullNames.length; i >= 0; i -= 1) {
+  for (let i = fullNames.length - 1; i >= 0; i--) {
     fullNames[i] = fullNames[i].replace("\n", "");
     if (fullNames[i] === TO_REMOVE) {
       fullNames.splice(i, 1);
@@ -22,10 +29,16 @@ function resolveNames(raw_names) {
   };
 }
 
+/**
+ *  Retrieves profileLinks and messageLinks of birthday profiles
+ *
+ * @param {*} raw_links - array of profile links in raw form
+ * @returns an object or arrys of profileLinks and messageLinks
+ */
 function resolveLinks(raw_links) {
   let profileLinks = [];
   const spaceSeparatedText = raw_links.split(" ");
-  for (let i = 0; i < spaceSeparatedText.length; i += 1) {
+  for (let i = 0; i < spaceSeparatedText.length; i++) {
     if (TO_MATCH.indexOf(spaceSeparatedText[i]) >= 0) {
       profileLinks.push(
         spaceSeparatedText[i += 1].replace(/href=/g, "").replace(/"/g, "")
